Add new directory and new mod ( Discovered how multiple files can communicate)
